refactor structure variation operators and selectors
