+++ qdax/core/emitters/selectors/abstract_selector.py
@@ -1,0 +1,18 @@
+import abc
+from typing import Tuple
+
+from qdax.core.containers.repertoire import Repertoire
+from qdax.core.emitters.emitter import EmitterState
+from qdax.types import Genotype, RNGKey
+
+
+class Selector(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def select(
+        self,
+        number_parents_to_select: int,
+        repertoire: Repertoire,
+        emitter_state: EmitterState,
+        random_key: RNGKey,
+    ) -> Tuple[Genotype, EmitterState, RNGKey]:
+        ...
